+++ packages/client/src/app/App.tsx
@@ -1,4 +1,4 @@
-import { Outlet, useLoaderData, useSearchParams } from 'react-router-dom'
+import { Outlet, useLoaderData } from 'react-router-dom'
 
 import { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
@@ -8,29 +8,13 @@ import { AppDispatch } from './store'
 import { User, setUser, setError } from '@processes/auth/model/authSlice'
 
 import ErrorBoundary from '@shared/components/ErrorBoundary'
-import { signinWithOauthToken } from '@processes/auth/api/authApi'
+import { useOAuthCodeHandle } from '@processes/auth'
 
 export const App = () => {
   const dispatch: AppDispatch = useDispatch()
   const userData: User | null | unknown = useLoaderData()
 
-  const [params, setParams] = useSearchParams()
-
-  const signin = async (code: string) => {
-    await signinWithOauthToken(
-      code,
-      `${window.location.protocol}//${window.location.host}`
-    )
-    setParams('')
-  }
-
-  useEffect(() => {
-    const oAuthCode = params.get('code')
-
-    if (oAuthCode) {
-      signin(oAuthCode)
-    }
-  }, [params])
+  useOAuthCodeHandle()
 
   useEffect(() => {
     if (userData) {
