+++ packages/client/src/processes/auth/ui/SigninForm.tsx
@@ -17,7 +17,7 @@ import { AppDispatch } from '@app/store'
 import { ROUTES } from '@shared/config/routes'
 import { useToast } from '@shared/hooks/use-toast'
 import { FormFieldWrapper } from './FormFieldWrapper'
-import { getOAuthServiceId } from '../api/authApi'
+import { OAuthYandexButton } from '@widgets/OAuthYandexButton/ui/OAuthYandexButton'
 
 const formSchema = z.object({
   login: validationRules.login,
@@ -53,24 +53,6 @@ const SigninForm: FC = () => {
     }
   }
 
-  const onSubmitOAuth = async () => {
-    try {
-      console.log('oauth')
-      const redirectUrl = 'http://localhost:3000'
-      const serviceId = await getOAuthServiceId(redirectUrl)
-
-      if (serviceId) {
-        const OAuthUrl = new URL('https://oauth.yandex.ru/authorize')
-        OAuthUrl.searchParams.set('response_type', 'code')
-        OAuthUrl.searchParams.set('client_id', serviceId)
-        OAuthUrl.searchParams.set('redirect_uri', redirectUrl)
-        location.href = OAuthUrl.toString()
-      }
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
@@ -83,10 +65,8 @@ const SigninForm: FC = () => {
         <Button className="w-full" type="submit">
           {loading ? 'Загрузка...' : 'Войти'}
         </Button>
-        <Button onClick={onSubmitOAuth} className="w-full" type="button">
-          Войти через Яндекс
-        </Button>
       </form>
+      <OAuthYandexButton />
     </Form>
   )
 }
