+++ packages/client/src/shared/lib/getCurrentHostWithPort.ts
@@ -1,0 +1,2 @@
+export const getCurrentHostWithPort = () =>
+  `${window.location.protocol}//${window.location.host}`
